Create pythonCode
CONTEST: In groups of 2 students, please try to develop a statistically valid predictor of CDRGLOB via using feature engineering. This means you need to work on the numeric values of the features (variables/columns) by performing multiplications and/or division operations. One such instance is BMI in which we divide weight by height square to get a good indicator of health (skinny, normal, obese, very obese etc). Thus, you should provide a mathematical formula and you should provide arguments and statistical analyses for the validation of your formula.

Here is an example that is just to illustrate how it can happen:

data["Tunas Formula"] = (data["NACCBMI"] * data["EDUC"]) / (data["HIPPOVOL"] * data["HIPPOVOL"] * data["INDEPEND"])

Evaluation metric for the contest: 1- You should provide the outputs for a correlation analysis between CDRGLOB and your formula's score. 2- You should also provide the group mean values for your formula's scores for each of 4 groups (CDRGLOB). 3- You should provide a visual illustration for the scores (of your formula) with respect to a variable (you should demonstrate the output in a scatter plot in which the cases should be colored with respect to CDRGLOB values).